Refactor and overhaul test_official
